Use a single policy for Transfers
Because creating incoming and outgoing transfers requires the same
permissions, we're creating a concern that we can share between the two
policies.
