+++ app/policies/incoming_transfer_policy.rb
@@ -1,0 +1,7 @@
+class IncomingTransferPolicy < ApplicationPolicy
+  include TransferPolicy
+
+  def target_activity
+    record.destination
+  end
+end
