+++ spec/policies/incoming_transfer_policy_spec.rb
@@ -1,0 +1,7 @@
+require "rails_helper"
+
+RSpec.describe IncomingTransferPolicy do
+  include_examples "transfer policy" do
+    let(:transfer) { create(:incoming_transfer, destination: activity) }
+  end
+end
